pyproject.toml
[project]
name = "langchain-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
]
